Changed from double to int

+++ src/TempConverter.java
@@ -1,11 +1,11 @@
 public class TempConverter {
 
-    public static double celsiusToFahrenheit(double celsius) {
+    public static int celsiusToFahrenheit(int celsius) {
         return (celsius * 9/5) + 32;
     }
 
     // Fahrenheit to Celsius conversion
-    public static double fahrenheitToCelsius(double fahrenheit) {
+    public static int fahrenheitToCelsius(int fahrenheit) {
         return (fahrenheit - 32) * 5/9;
     }
 }
